access_token刷新修改分布式锁下的双重检验机制

+++ wechat-core/src/main/java/org/redrock/wechatcore/bean/Token.java
@@ -11,7 +11,7 @@ public class Token extends WechatError{
     String accessToken;
 
     @JsonAlias("expires_in")
-    int expiresIn;
+    long expiresIn;
 
     @JsonAlias("refresh_token")
     String refreshToken;
